--- a/converter/src/main/java/tech/allegro/schema/json2avro/converter/JsonGenericRecordReader.java
+++ b/converter/src/main/java/tech/allegro/schema/json2avro/converter/JsonGenericRecordReader.java
@@ -12,20 +12,18 @@ import java.util.ArrayDeque;
 import java.util.Deque;
 import java.util.List;
 import java.util.Map;
-import java.util.Spliterator;
-import java.util.Spliterators;
-import java.util.stream.Collectors;
-import java.util.stream.StreamSupport;
+import java.util.function.Function;
 
-import static java.lang.String.format;
 import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.*;
+import static tech.allegro.schema.json2avro.converter.AvroTypeExceptions.*;
 
 public class JsonGenericRecordReader {
-    private ObjectMapper mapper;
+    private static final Object INCOMPATIBLE = new Object();
+    private final ObjectMapper mapper;
 
     public JsonGenericRecordReader() {
-        this.mapper = new ObjectMapper();
+        this(new ObjectMapper());
     }
 
     public JsonGenericRecordReader(ObjectMapper mapper) {
@@ -54,12 +52,12 @@ public class JsonGenericRecordReader {
             GenericRecordBuilder record = new GenericRecordBuilder(schema);
             json.entrySet().forEach(entry ->
                     ofNullable(schema.getField(entry.getKey()))
-                            .ifPresent(field -> record.set(field, read(field, field.schema(), entry.getValue(), path))));
+                            .ifPresent(field -> record.set(field, read(field, field.schema(), entry.getValue(), path, false))));
             return record.build();
     }
 
     @SuppressWarnings("unchecked")
-    private Object read(Schema.Field field, Schema schema, Object value, Deque<String> path) {
+    private Object read(Schema.Field field, Schema schema, Object value, Deque<String> path, boolean silently) {
         boolean pushed = !field.name().equals(path.peek());
         if(pushed) {
             path.push(field.name());
@@ -67,18 +65,18 @@ public class JsonGenericRecordReader {
         Object result;
 
         switch (schema.getType()) {
-            case RECORD:  result = readRecord(ensureType(value, Map.class, path), schema, path); break;
-            case ARRAY:   result =  readArray(field, schema, ensureType(value, List.class, path), path); break;
-            case MAP:     result =  readMap(field, schema, ensureType(value, Map.class, path), path); break;
-            case UNION:   result =  readUnion(field, schema, value, path); break;
-            case INT:     result =  ensureType(value, Number.class, path).intValue(); break;
-            case LONG:    result =  ensureType(value, Number.class, path).longValue(); break;
-            case FLOAT:   result =  ensureType(value, Number.class, path).floatValue(); break;
-            case DOUBLE:  result =  ensureType(value, Number.class, path).doubleValue(); break;
-            case BOOLEAN: result =  ensureType(value, Boolean.class, path); break;
-            case ENUM:    result =  ensureEnum(schema, ensureType(value, String.class, path), path); break;
-            case STRING:  result =  ensureType(value, String.class, path); break;
-            case NULL:    result =  ensureNull(value, path); break;
+            case RECORD:  result = onValidType(value, Map.class, path, silently, map -> readRecord(map, schema, path)); break;
+            case ARRAY:   result = onValidType(value, List.class, path, silently, list -> readArray(field, schema, list, path)); break;
+            case MAP:     result = onValidType(value, Map.class, path, silently, map -> readMap(field, schema, map, path)); break;
+            case UNION:   result = readUnion(field, schema, value, path); break;
+            case INT:     result = onValidNumber(value, path, silently, Number::intValue); break;
+            case LONG:    result = onValidNumber(value, path, silently, Number::longValue); break;
+            case FLOAT:   result = onValidNumber(value, path, silently, Number::floatValue); break;
+            case DOUBLE:  result = onValidNumber(value, path, silently, Number::doubleValue); break;
+            case BOOLEAN: result = onValidType(value, Boolean.class, path, silently, bool -> bool); break;
+            case ENUM:    result = onValidType(value, String.class, path, silently, string -> ensureEnum(schema, string, path)); break;
+            case STRING:  result = onValidType(value, String.class, path, silently, string -> string); break;
+            case NULL:    result = value == null ? value : INCOMPATIBLE; break;
             default: throw new AvroTypeException("Unsupported type: " + field.schema().getType());
         }
 
@@ -89,28 +87,34 @@ public class JsonGenericRecordReader {
     }
 
     private List<Object> readArray(Schema.Field field, Schema schema, List<Object> items, Deque<String> path) {
-        return items.stream().map(item -> read(field, schema.getElementType(), item, path)).collect(toList());
+        return items.stream().map(item -> read(field, schema.getElementType(), item, path, false)).collect(toList());
     }
 
     private Map<String, Object> readMap(Schema.Field field, Schema schema, Map<String, Object> map, Deque<String> path) {
         return map.entrySet()
                 .stream()
-                .collect(toMap(Map.Entry::getKey, entry -> read(field, schema.getValueType(), entry.getValue(), path)));
+                .collect(toMap(Map.Entry::getKey, entry -> read(field, schema.getValueType(), entry.getValue(), path, false)));
     }
-
 
     private Object readUnion(Schema.Field field, Schema schema, Object value, Deque<String> path) {
         List<Schema> types = schema.getTypes();
         for (Schema type : types) {
             try {
-                return read(field, type, value, path);
-            } catch (AvroRuntimeException ex) {
+                Object nestedValue = read(field, type, value, path, true);
+                if (nestedValue == INCOMPATIBLE) {
+                    continue;
+                } else {
+                    return nestedValue;
+                }
+            } catch (AvroRuntimeException e) {
+                // thrown only for union of more complex types like records
                 continue;
             }
         }
-        throw new AvroTypeException(format("Could not evaluate union, field %s is expected to be one of these: %s. " +
-                "If this is a complex type, check if offending field: %s adheres to schema.",
-                field.name(), types.stream().map(Schema::getType).map(Object::toString).collect(joining(",")), path(path)));
+        throw unionException(
+                field.name(),
+                types.stream().map(Schema::getType).map(Object::toString).collect(joining(", ")),
+                path);
     }
 
     private Object ensureEnum(Schema schema, Object value, Deque<String> path) {
@@ -118,28 +122,25 @@ public class JsonGenericRecordReader {
         if(symbols.contains(value)){
            return value;
         }
-        String knownSymbols = symbols.stream().map(String::valueOf).collect(Collectors.joining(", "));
-        throw new AvroTypeException(format("Field %s is expected to be of enum type and be one of %s", path(path), knownSymbols));
+        throw enumException(path, symbols.stream().map(String::valueOf).collect(joining(", ")));
     }
 
-    @SuppressWarnings("unchecked")
-    private <T> T ensureType(Object value, Class<T> type, Deque<String> path) {
+    public <T> Object onValidType(Object value, Class<T> type, Deque<String> path, boolean silently, Function<T, Object> function)
+            throws AvroTypeException {
+
         if (type.isInstance(value)) {
-            return (T) value;
+            return function.apply((T) value);
+        } else {
+            if (silently) {
+                return INCOMPATIBLE;
+            } else {
+                throw typeException(path, type.getTypeName());
+            }
         }
-        throw new AvroTypeException(format("Field %s is expected to be of %s type.", path(path), type.getName()));
     }
 
-    private Object ensureNull(Object o, Deque<String> path) {
-        if (o != null) {
-            throw new AvroTypeException(format("Field %s was expected to be null.", path(path)));
-        }
-        return null;
-    }
-
-    private String path(Deque<String> path) {
-        return StreamSupport.stream(Spliterators.spliteratorUnknownSize(path.descendingIterator(), Spliterator.ORDERED), false)
-            .map(Object::toString).collect(joining("."));
+    public Object onValidNumber(Object value, Deque<String> path, boolean silently, Function<Number, Object> function) {
+        return onValidType(value, Number.class, path, silently, function);
     }
 }