This PR speeds up conversion from json to avro through elimination of:
* exceptions from normal flow 
* usage of String.format()

# Before optimization
```
Benchmark                                                                                Mode     Cnt        Score    Error  Units
JsonAvroConverterBenchmark.conversionThroughputForCompleteMessage                       thrpt       2   269449.055           ops/s
JsonAvroConverterBenchmark.conversionThroughputForMessageWithNotProvidedOptionalField   thrpt       2  2448377.702           ops/s
JsonAvroConverterBenchmark.conversionLatencyForCompleteMessage                         sample  362647     3534.198 ± 68.884  ns/op
JsonAvroConverterBenchmark.conversionLatencyForMessageWithNotProvidedOptionalField     sample  381981      461.864 ± 14.318  ns/op
```

# After optimization
```
Benchmark                                                                                Mode     Cnt        Score    Error  Units
JsonAvroConverterBenchmark.conversionThroughputForCompleteMessage                       thrpt       2  1895774.508           ops/s
JsonAvroConverterBenchmark.conversionThroughputForMessageWithNotProvidedOptionalField   thrpt       2  2397447.209           ops/s
JsonAvroConverterBenchmark.conversionLatencyForCompleteMessage                         sample  287009      591.431 ±  7.370  ns/op
JsonAvroConverterBenchmark.conversionLatencyForMessageWithNotProvidedOptionalField     sample  367727      480.230 ± 15.620  ns/op
```

# Root causes and solutions
```
Benchmark                                                       Mode     Cnt     Score    Error  Units
LanguageBenchmark.benchExceptionThrow                          sample  246564  1324.748 ± 21.692  ns/op
LanguageBenchmark.benchObjectReturn                            sample  322717    50.053 ±  0.622  ns/op
LanguageBenchmark.benchStringFormat                            sample  296719  1112.386 ± 24.829  ns/op
LanguageBenchmark.benchStringsConcatenationViaBuilder          sample  366216    69.523 ±  4.435  ns/op
```